+++ src/app/modules/project/overview/overview.component.ts
@@ -251,9 +251,11 @@ export class OverviewComponent implements OnInit {
   /**
    * Triggers on project click in the list.
    * @param id project id.
+   * @param name project name
    */
-  public onClickProject(id: number): void {
-    this.router.navigate([`/project/details/${id}`]);
+  public onClickProject(id: number, name: string): void {
+    name = name.split(' ').join('-');
+    this.router.navigate([`/project/details/${id}-${name}`]);
   }
 
   /**